face-recognitions/requirements.txt
face_recognition==1.3.0
Flask==3.0.3
keras==2.12.0
Keras_Preprocessing==1.1.2
mtcnn==0.1.1
numpy==1.24.2
opencv_contrib_python==4.9.0.80
opencv_python==4.7.0.68
Pillow==10.3.0
h5py==2.10.0
imutils==0.5.3
